requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.5
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.1
Django==1.11.20
django-crispy-forms==1.9.0
django-phonenumber-field==4.0.0
django-redis==4.11.0
djangorestframework==3.11.0
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumberslite==8.11.5
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
python-http-client==3.2.6
pytz==2019.3
redis==3.4.1
sendgrid==6.1.3
six==1.14.0
typed-ast==1.4.1
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
